+++ engine/src/application.h
@@ -37,7 +37,7 @@ namespace emerald {
 
 		uint32_t getWidth() const;
 		uint32_t getHeight() const;
-		double getTime() const;
+		float getTime() const;
 		Ref<Window>& getWindow() { return m_mainWindow; }
 
 		template<typename F>
